docker_app/requirements.txt
beautifulsoup4
pyyaml
requests